pyproject.toml
[project]
name = "qobo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy==1.26.4",
    "pyyaml>=6.0.3",
    "torch==2.2.2",
    "tqdm>=4.67.3",
]
